dotstrap: Improve Sublime Text configuration
Ask users if they want to actually configure Sublime Text, since you
don't need that on e.g. a server environment. Also install the package
manager preemptively, since Sublime Text is really ugly initially due to
missing packages and this is faster that waiting for Sublime Text to
download it.

Signed-off-by: Stephen Finucane <stephen@that.guru>